pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "support-triage-openenv"
version = "0.1.0"
description = "OpenEnv: support desk ticket triage with graded tasks"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn[standard]>=0.24.0",
]

[project.optional-dependencies]
inference = ["openai>=1.40.0"]
dev = ["pytest>=8.0.0", "httpx>=0.27.0", "openai>=1.40.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["support_triage_env*", "server*"]

[project.scripts]
server = "support_triage_env.server.app:main"
support-triage-server = "support_triage_env.server.app:main"
